Hotfix for the "print failed (last_cit) but
When using the diogenes cli, somehow a error message clutters the
output:

  Print failed (first_cit)

By making this error message only printed in debug mode, I hotfixed
this issue. (It is really necessary to print these messages in normal
mode?) It seems that the problem lies somewhere in
Diogenes::Browser->maybe_use_cit.